requirements.txt
asgiref==3.4.1
black==21.11b1
click==8.0.3
Django==3.2.9
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
python-decouple==3.5
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==4.0.1
